backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
psycopg2-binary==2.9.9
pydantic==2.6.1